Clean code and doc

+++ wce/pipeline.py
@@ -1,3 +1,11 @@
+"""Pipeline
+
+Module implementing two commands for the WCE pipeline:
+
+    * train - train a WCE model.
+    * predict - predict the word counts of audio files.
+"""
+
 import os, sys, glob
 import csv
 import shutil
@@ -7,23 +15,40 @@ from .envelope_estimation.feature_extraction import FeatureExtraction
 from .envelope_estimation.batch import Batch
 from .envelope_estimation.envelope_estimator import EnvelopeEstimator
 from .word_count_estimation.annotations_processing import process_annotations
-from .word_count_estimation.speech_extractor import extract_speech, \
+from .word_count_estimation.speech_extractor import extract_speech_from_dir, \
                                                     retrieve_files_word_counts
 from .word_count_estimation.word_count_estimator import WordCountEstimator
 
 
-def train(audio_dir, annotations_dir, rttm_dir, sad_name, data_processing_config,
-          env_model_path, wce_model_path, selcha_script):
+def train(audio_dir, annotations_dir, data_processing_config, env_model_path,
+          wce_model_path, selcha_script, rttm_dir=None, sad_name=None):
     """
-    TODO
+    Pipeline to train a WCE model.
+
+    Parameters
+    ----------
+    audio_dir : str
+        Directory where the audio files are stored.
+    annotations_dir : str
+        Directory where the annotation files are stored.
+    data_processing_config : str
+        Path to the data processing config file.
+    env_model_path : str
+        Path to the syllable envelope estimator model file.
+    wce_model_path : str
+        Path to the word count estimator model file.
+    rttm_dir : str, optional
+        Directory where the SAD .rttm files are stored.
+        Defaults to None.
+    sad_name : str, optional
+        Name of the SAD algorithm used.
+        Defaults to None.
     """
 
     if not os.path.exists(audio_dir):
         raise IOError("Audio directory does not exist.")
     if not os.path.exists(annotations_dir):
         raise IOError("Annotation directory does not exist.")
-    if not os.path.exists(rttm_dir):
-        raise IOError("SAD directory does not exist.")
     if not os.path.exists(data_processing_config):
         raise IOError("Data processing config file does not exist.")
     if not os.path.exists(env_model_path):
@@ -44,9 +69,17 @@ def train(audio_dir, annotations_dir, rttm_dir, sad_name, data_processing_config
             process_annotations(audio_dir, annotations_dir, rttm_dir,
                                 sad_name, selcha_script)
 
-    wce_model.alpha = alpha
     audio_files = wav_list
-    target_counts = tot_seg_words
+    if rttm_dir:
+        if not os.path.exists(rttm_dir):
+            raise IOError("SAD directory does not exist.")
+        if not sad_name:
+            sys.exit("SAD name missing.")
+        wce_model.alpha = alpha
+        target_counts = tot_seg_words
+    else:
+        wce_model.alpha = 1
+        target_counts = tot_files_words
 
     print("Extracting features from data.")
     feature_list = []
@@ -70,18 +103,40 @@ def train(audio_dir, annotations_dir, rttm_dir, sad_name, data_processing_config
     print("Training WCE model.")
     wce_model.train(envelopes, target_counts, model_file=wce_model_path)
 
-    chunks_dir = os.path.dirname(audio_files[0])
-    shutil.rmtree(chunks_dir)
+    if rttm_dir:
+        chunks_dir = os.path.dirname(audio_files[0])
+        shutil.rmtree(chunks_dir)
 
     print("Training ended successfully.")
     print("WCE model saved at: {}".format(wce_model_path))
     wce_model.summary()
 
 
-def predict(audio_dir, results_path, data_processing_config,
-            env_model_path, wce_model_path, rttm_dir=None, sad_name=None):
+def predict(audio_dir, results_path, data_processing_config, env_model_path,
+            wce_model_path, rttm_dir=None, sad_name=None):
     """
-    TODO
+    Pipeline to predict the word count of audio files.
+
+    Parameters
+    ----------
+    audio_dir : str
+        Directory where the audio files are stored.
+    results_path : 
+        Path to the word count output .csv file.
+    data_processing_config : str
+        Path to the data processing config file.
+    env_model_path : str
+        Path to the syllable envelope estimator model file.
+    wce_model_path : str
+        Path to the word count estimator model file.
+    selcha_script : str
+        Path to the selcha script.
+    rttm_dir : str, optional
+        Directory where the SAD .rttm files are stored.
+        Defaults to None.
+    sad_name : str, optional
+        Name of the SAD algorithm used.
+        Defaults to None.
     """
 
     if not os.path.exists(audio_dir):
@@ -102,15 +157,13 @@ def predict(audio_dir, results_path, data_processing_config,
     wce_model = WordCountEstimator()
     wce_model.load_model(wce_model_path)
 
-    if rttm_dir and sad_name:
+    if rttm_dir:
         if not os.path.exists(rttm_dir):
             raise IOError("SAD directory does not exist.")
+        if not sad_name:
+            sys.exit("SAD name missing.")
         print("Extracting speech segments from audio files.")
-        audio_files = extract_speech(audio_dir, rttm_dir, sad_name)
-    elif rttm_dir:
-        sys.exit("SAD name missing.")
-    elif sad_name:
-        sys.exit("rttm directory missing.")
+        audio_files = extract_speech_from_dir(audio_dir, rttm_dir, sad_name)
     else:
         audio_files = glob.glob(os.path.join(audio_dir, "*.wav"))
